+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/nodeviews/mediaNodeUpdater.spec.ts
@@ -14,7 +14,6 @@ import {
   MediaNodeUpdater,
   MediaNodeUpdaterProps,
 } from '../../../../../plugins/media/nodeviews/mediaNodeUpdater';
-import * as mediaCommon from '../../../../../plugins/media/utils/media-common';
 import { MediaProvider } from '../../../../../plugins/media/pm-plugins/main';
 
 describe('MediaNodeUpdater', () => {
@@ -25,12 +24,11 @@ describe('MediaNodeUpdater', () => {
     jest
       .spyOn(MediaClientModule, 'getMediaClient')
       .mockReturnValue(mediaClient);
-    jest.spyOn(commands, 'updateAllMediaNodesAttrs').mockReturnValue(() => {});
-    jest.spyOn(commands, 'updateMediaNodeAttrs').mockReturnValue(() => {});
-    jest.spyOn(commands, 'replaceExternalMedia').mockReturnValue(() => {});
     jest
-      .spyOn(mediaCommon, 'getViewMediaClientConfigFromMediaProvider')
-      .mockReturnValue(getDefaultMediaClientConfig());
+      .spyOn(commands, 'updateAllMediaNodesAttrs')
+      .mockReturnValue(() => true);
+    jest.spyOn(commands, 'updateMediaNodeAttrs').mockReturnValue(() => true);
+    jest.spyOn(commands, 'replaceExternalMedia').mockReturnValue(() => true);
 
     const contextIdentifierProvider: Promise<
       ContextIdentifierProvider