+++ packages/media/media-card/src/utils/__tests__/unit/getElementDimension.tsx
@@ -13,7 +13,11 @@ describe('getElementDimension', () => {
         };
       },
     };
-    (ReactDOM.findDOMNode as jest.Mock<void>).mockReturnValue(element);
+    ((ReactDOM.findDOMNode as (
+      instance: React.Component<any, {}, any> | Element | null | undefined,
+    ) => Element | Text | null) as jest.Mock<Element>).mockReturnValue(
+      element as Element,
+    );
     const dummyComponent = <div /> as any; // This casting is done to avoid having to create a new React class here
     const width = getElementDimension(dummyComponent, 'width');
     const height = getElementDimension(dummyComponent, 'height');