+++ packages/media/media-client/src/__tests__/unit/fileFetcher.spec.ts
@@ -1,10 +1,11 @@
-import { Auth, AuthProvider, getFileStreamsCache } from '@atlaskit/media-core';
+import { Auth, AuthProvider } from '@atlaskit/media-core';
 import {
   ResponseFileItem,
   MediaFile,
   RECENTS_COLLECTION,
   globalMediaEventEmitter,
   MediaViewedEventPayload,
+  getFileStreamsCache,
 } from '../..';
 import * as MediaClientModule from '../..';
 import uuid from 'uuid';
@@ -197,7 +198,7 @@ describe('FileFetcher', () => {
 
       fileFetcher.getFileState('invalid-id').subscribe({
         error(error) {
-          expect(error).toEqual('invalid-id is not a valid file id');
+          expect(error).toEqual('invalid id was passed to getFileState');
           done();
         },
       });
@@ -250,6 +251,8 @@ describe('FileFetcher', () => {
 
   describe('copyFile', () => {
     it('should call mediaStore.copyFileWithToken', async () => {
+      // @ts-ignore This violated type definition upgrade of @types/jest to v24.0.18 & ts-jest v24.1.0.
+      //See BUILDTOOLS-210-clean: https://bitbucket.org/atlassian/atlaskit-mk-2/pull-requests/7178/buildtools-210-clean/diff
       const MediaStoreSpy = jest.spyOn(MediaClientModule, 'MediaStore');
       const { items, fileFetcher } = setup();
       const copyFileWithTokenMock = jest.fn().mockResolvedValue({ data: {} });
@@ -296,6 +299,8 @@ describe('FileFetcher', () => {
     });
 
     it('should populate cache with the copied file', async () => {
+      // @ts-ignore This violated type definition upgrade of @types/jest to v24.0.18 & ts-jest v24.1.0.
+      //See BUILDTOOLS-210-clean: https://bitbucket.org/atlassian/atlaskit-mk-2/pull-requests/7178/buildtools-210-clean/diff
       const MediaStoreSpy = jest.spyOn(MediaClientModule, 'MediaStore');
       const copiedFile: MediaFile = {
         id: 'copied-file-id',