+++ packages/media/media-picker/src/components/popup.tsx
@@ -41,7 +41,7 @@ export class PopupImpl extends UploadComponent<PopupUploadEventPayloadMap>
     super();
     this.proxyReactContext = proxyReactContext;
 
-    const { userAuthProvider, cacheSize } = tenantMediaClient.config;
+    const { userAuthProvider } = tenantMediaClient.config;
     if (!userAuthProvider) {
       throw new Error(
         'When using Popup media picker userAuthProvider must be provided in the context',
@@ -49,7 +49,6 @@ export class PopupImpl extends UploadComponent<PopupUploadEventPayloadMap>
     }
 
     const userMediaClient = new MediaClient({
-      cacheSize,
       authProvider: userAuthProvider,
     });
     const tenantUploadParams = {