+++ packages/core/dynamic-table/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/dynamic-table
 
+## 13.5.0
+
+### Minor Changes
+
+- [minor][19e5d77c84](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/19e5d77c84):
+
+  Adding an optional prop `testId` that will set the attribute value `data-testid`. It will help products to write better integration and end to end tests.
+
 ## 13.4.0
 
 ### Minor Changes